requirements.txt
daily-python
requests
python-dotenv
vapi-python 
pyaudio